extract contributed commands collection
